Create experiment package.
1. Rename starfish/experiment.py to starfish/experiment/__init__.py
2. Move examples/support/__init__.py into starfish/experiment/
3. Rename examples/build_sample_experiment.py to starfish/experiment/builder/cli.py
