+++ starfish/experiment/builder/__init__.py
@@ -1,10 +1,7 @@
 import json
 import os
-from io import BytesIO
-from typing import IO, Tuple
+from typing import Tuple
 
-import numpy as np
-from skimage.io import imsave
 from slicedimage import (
     Collection,
     ImageFormat,
@@ -14,6 +11,7 @@ from slicedimage import (
 )
 
 from starfish.types import Coordinates, Indices
+from .imagedata import FetchedImage, ImageFetcher, RandomNoiseImage, RandomNoiseImageFetcher
 
 
 AUX_IMAGE_NAMES = {
@@ -46,66 +44,6 @@ def fov_path_generator(parent_toc_path, toc_name):
         os.path.dirname(parent_toc_path),
         "{}-{}.json".format(toc_basename, toc_name),
     )
-
-
-class FetchedImage:
-    """
-    This is the contract for providing the data for constructing a :class:`slicedimage.Tile`.
-    """
-    @property
-    def shape(self) -> Tuple[int, ...]:
-        raise NotImplementedError()
-
-    @property
-    def format(self) -> ImageFormat:
-        raise NotImplementedError()
-
-    @property
-    def image_data_handle(self) -> IO:
-        raise NotImplementedError()
-
-
-class ImageFetcher:
-    """
-    This is the contract for providing the image data for constructing a
-    :class:`slicedimage.Collection`.
-    """
-    def get_image(self, fov: int, hyb: int, ch: int, z: int) -> FetchedImage:
-        """
-        Given fov, hyb, ch, and z, return an instance of a :class:`.FetchedImage` that can be
-        queried for the image data.
-        """
-        raise NotImplementedError()
-
-
-class RandomNoiseImage(FetchedImage):
-    """
-    This is a simple implementation of :class:`.FetchedImage` that simply regenerates random data
-    for the image.
-    """
-    @property
-    def shape(self) -> Tuple[int, ...]:
-        return 1536, 1024
-
-    @property
-    def format(self) -> ImageFormat:
-        return ImageFormat.TIFF
-
-    def image_data_handle(self) -> IO:
-        arr = np.random.randint(0, 256, size=self.shape, dtype=np.uint8)
-        output = BytesIO()
-        imsave(output, arr, plugin="tifffile")
-        output.seek(0)
-        return output
-
-
-class RandomNoiseImageFetcher(ImageFetcher):
-    """
-    This is a simple implementation of :class:`.ImageFetcher` that simply returns a
-    :class:`.RandomNoiseImage` for every fov, hyb, ch, z combination.
-    """
-    def get_image(self, fov: int, hyb: int, ch: int, z: int) -> FetchedImage:
-        return RandomNoiseImage()
 
 
 def build_image(
